pyproject.toml
[project]
name = "binance-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.8",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=0.1.0",
    "python-binance>=1.0.34",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
